requirements_simple.txt
# AI Gold Trading Bot - Python Dependencies (Simple Version)
# For Python 3.12 compatibility - no version locks
# Installation: pip install -r requirements_simple.txt

# ============================================================================
# Core AI/ML Libraries
# ============================================================================
tensorflow
xgboost
scikit-learn
transformers
torch
torchvision

# ============================================================================
# Trading Platform
# ============================================================================
MetaTrader5

# ============================================================================
# Data Science
# ============================================================================
pandas
numpy
matplotlib
seaborn
plotly
mplfinance

# ============================================================================
# Technical Indicators
# ============================================================================
pandas-ta
# TA-Lib  # Install separately - see instructions below

# ============================================================================
# Data Collection
# ============================================================================
beautifulsoup4
selenium
requests
yfinance
newsapi-python

# ============================================================================
# Database
# ============================================================================
sqlalchemy

# ============================================================================
# Utilities
# ============================================================================
python-telegram-bot
schedule
python-dotenv
loguru
tqdm
joblib

# ============================================================================
# Image Processing
# ============================================================================
Pillow
opencv-python

# ============================================================================
# Date/Time
# ============================================================================
pytz
python-dateutil

# ============================================================================
# Data Validation
# ============================================================================
pydantic
typing-extensions

# ============================================================================
# Testing
# ============================================================================
pytest
pytest-cov
pytest-mock

# ============================================================================
# Code Quality
# ============================================================================
black
flake8
pylint
mypy

# ============================================================================
# Jupyter Notebook
# ============================================================================
jupyter
ipykernel
notebook

# ============================================================================
# Performance
# ============================================================================
numba

# ============================================================================
# Configuration
# ============================================================================
pyyaml
toml

# ============================================================================
# Web Framework (optional)
# ============================================================================
flask

# ============================================================================
# Additional ML
# ============================================================================
lightgbm
statsmodels
scipy

# ============================================================================
# NLP
# ============================================================================
nltk
textblob

# ============================================================================
# Visualization
# ============================================================================
bokeh

# ============================================================================
# INSTALLATION INSTRUCTIONS
# ============================================================================
#
# 1. Install all packages:
#    pip install -r requirements_simple.txt
#
# 2. Install TA-Lib separately:
#
#    Windows:
#      Download from: https://github.com/cgohlke/talib-build/releases
#      Choose: TA_Lib-0.4.32-cp312-cp312-win_amd64.whl (for Python 3.12)
#      Install: pip install TA_Lib-0.4.32-cp312-cp312-win_amd64.whl
#
#    Mac:
#      brew install ta-lib
#      pip install TA-Lib
#
#    Linux:
#      sudo apt-get install ta-lib
#      pip install TA-Lib
#
# 3. Verify installation:
#    python test_setup.py
#
# ============================================================================
